Feat: update

+++ src/pages/auth/Confirm.jsx
@@ -5,27 +5,25 @@ import { NavLink, useNavigate } from 'react-router-dom'
 import { useAuthContext } from '../../context/authContext'
 import { endpoint } from '../../utils/Endpoints'
 
-const Login = () => {
+const Confirm = () => {
     const navigate = useNavigate()
 
     const [state, setState] = useState({
-        details: { email: '', password: '' }
+        details: { email: '', verification_code: '' }
     })
 
-    const SubmitLogin = (e) => {
+    const submitConfirm = (e) => {
         e.preventDefault();
 
-        axios.post(endpoint.login, state.details)
+        axios.post(endpoint.confirm, state.details)
         .then(res=>{
             console.log(res.data)
             localStorage.setItem('user', JSON.stringify({
-                id: res.data.user_data.id,
                 email: res.data.user_data.Email,
-                first_name:res.data.user_data.First_name,
-                last_name:res.data.user_data.Last_name,
-                username:res.data.user_data.Username,
+                code:res.data.user_data.First_name,
+                
             }))
-            localStorage.setItem('token', res.data.tokens.access_token)
+            localStorage.setItem('token', res.data.access_token)
             localStorage.setItem('isLogin', 'true')
 
             navigate('/', {replace: true})
@@ -36,27 +34,17 @@ const Login = () => {
     }
 
 
-    return (
-        <div className='w-full relative h-[100vh] overflow-hidden'>
+  return (
+    <div className='w-full relative h-[100vh] overflow-hidden'>
             <div className='bgLogin w-full h-full flex text-white flex-col justify-center items-center z-10'>
                 <div className='max-w-[30rem] w-[40%]  mx-auto  text-white overflow-auto'>
                     <div className='w-full h-full'>
                         <div className='w-full text-center pb-4'>
-                            <h2 className='text-[3rem] capitalize font-semibold'>welcome back!</h2>
+                            <h2 className='text-[3rem] capitalize font-semibold'>Confirm mail</h2>
                             <p className='text-[1.1rem]'>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Explicabo, eius.</p>
                         </div>
                         <section className="w-[80%] mx-auto py-[2rem]">
-                            <form onSubmit={(e)=>SubmitLogin(e)} className='w-full flex flex-col justify-center items-center'>
-                                <div className='w-full mb-8 flex items-center gap-[0.7rem] py-[0.2rem] px-[1rem] border border-white'>
-                                    <div>
-                                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-                                            <path stroke-linecap="round" stroke-linejoin="round" d="M17.982 18.725A7.488 7.488 0 0 0 12 15.75a7.488 7.488 0 0 0-5.982 2.975m11.963 0a9 9 0 1 0-11.963 0m11.963 0A8.966 8.966 0 0 1 12 21a8.966 8.966 0 0 1-5.982-2.275M15 9.75a3 3 0 1 1-6 0 3 3 0 0 1 6 0Z" />
-                                        </svg>
-                                    </div>
-                                    <div className='text-[2rem] font-light'>|</div>
-
-                                    <input name='username' value={state.details.username} onChange={(e)=>setState({...state, details: {...state.details, username: e.target.value}})} type="text" className="w-[70%] text-[1.1rem] placeholder:text-white bg-transparent border-none outline-none" placeholder='Enter your username' />
-                                </div>
+                            <form onSubmit={(e)=>submitConfirm(e)} className='w-full flex flex-col justify-center items-center'>
                                 <div className='w-full mb-8 flex items-center gap-[0.7rem] py-[0.2rem] px-[1rem] border border-white'>
                                     <div>
                                         <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-8 h-8">
@@ -74,7 +62,7 @@ const Login = () => {
                                         </svg>
                                     </div>
                                     <div className='text-[2rem] font-light'>|</div>
-                                    <input name='password' value={state.details.password} onChange={(e)=>setState({...state, details: {...state.details, password: e.target.value}})} type="password" className="w-[70%] text-[1.1rem] placeholder:text-white bg-transparent border-none outline-none" placeholder='Enter your password' />
+                                    <input name='verification_code' value={state.details.verification_code} onChange={(e)=>setState({...state, details: {...state.details, verification_code: e.target.value}})} type="text" className="w-[70%] text-[1.1rem] placeholder:text-white bg-transparent border-none outline-none" placeholder='Enter your code' />
                                 </div>
 
                                 <div className='w-full flex justify-between items-center mb-10'>
@@ -85,14 +73,14 @@ const Login = () => {
                                     <div className='text-[1rem]'>Don't have an account? <NavLink to='/signup'>Signup</NavLink> </div>
                                 </div>
 
-                                <button className='py-[0.5rem] px-[4rem] text-[1.2rem] font-medium uppercase border border-white'>sign in </button>
+                                <button className='py-[0.5rem] px-[4rem] text-[1.2rem] font-medium uppercase border border-white'>Confirm</button>
                             </form>
                         </section>
                     </div>
                 </div>
             </div>
         </div>
-    )
+  )
 }
 
-export default Login
+export default Confirm
